fix: [AXIMST-408,409] fix tab title, learn more link blank, rename files, update translations
