+++ src/certificates/Certificates.jsx
@@ -1,8 +1,8 @@
 import PropTypes from 'prop-types';
 
-import MainLayout from './layout';
-import WithoutModes from './without-modes';
-import WithModesWithoutCertificates from './with-modes-without-certificates';
+import MainLayout from './layout/MainLayout';
+import WithoutModes from './without-modes/WithoutModes';
+import WithModesWithoutCertificates from './with-modes-without-certificates/WithModesWithoutCertificates';
 import useCertificates from './hooks/useCertificates';
 
 const Certificates = ({ courseId }) => {
