+++ src/certificates/with-modes-without-certificates/WithModesWithoutCertificates.jsx
@@ -3,7 +3,7 @@ import { useIntl } from '@edx/frontend-platform/i18n';
 import { Button, Stack, Card } from '@edx/paragon';
 import { Add as AddIcon } from '@edx/paragon/icons';
 
-import MainLayout from '../layout';
+import MainLayout from '../layout/MainLayout';
 import messages from '../messages';
 
 const WithModesWithoutCertificates = ({ courseId }) => {
