+++ src/certificates/without-modes/WithoutModes.test.jsx
@@ -5,7 +5,7 @@ import { AppProvider } from '@edx/frontend-platform/react';
 
 import initializeStore from '../../store';
 import messages from '../messages';
-import WithoutModes from '.';
+import WithoutModes from './WithoutModes';
 
 const courseId = 'course-123';
 let store;
