+++ src/i18n/messages/zh_CN.json
@@ -993,19 +993,20 @@
   "course-authoring.course-unit.add.component.title": "Add a new component",
   "course-authoring.course-unit.add.component.button.text": "Add Component:",
   "course-authoring.certificates.heading.title": "Certificates",
+  "course-authoring.certificates.heading.title.tab.text": "Course Certificates",
   "course-authoring.certificates.heading.subtitle": "Settings",
   "course-authoring.certificates.heading.action.button.preview": "Preview certificate",
   "course-authoring.certificates.heading.action.button.deactivate": "Deactivate",
   "course-authoring.certificates.nocertificate.text": "You haven't added any certificates to this course yet.",
   "course-authoring.certificates.setup.certificate.button": "Add your first certificate",
   "course-authoring.certificates.without.modes.text": "This course does not use a mode that offers certificates.",
-  "course-authoring.certificates.sidebar.about.title": "Working with Certificates",
+  "course-authoring.certificates.sidebar.about.title": "Working with certificates",
   "course-authoring.certificates.sidebar.about.description-1": "Specify a course title to use on the certificate if the course's official title is too long to be displayed well.",
   "course-authoring.certificates.sidebar.about.description-2": "For verified certificates, specify between one and four signatories and upload the associated images. To edit or delete a certificate before it is activated, hover over the top right corner of the form and select {strongText} or the delete icon.",
   "course-authoring.certificates.sidebar.about.description-2.strong": "Edit",
   "course-authoring.certificates.sidebar.about.description-3": "To view a sample certificate, choose a course mode and select {strongText}.",
-  "course-authoring.certificates.sidebar.about.description-3.strong": "Preview Certificate",
-  "course-authoring.certificates.sidebar.about2.title": "Issuing Certificates to Learners",
+  "course-authoring.certificates.sidebar.about.description-3.strong": "Preview certificate",
+  "course-authoring.certificates.sidebar.about2.title": "Issuing certificates to learners",
   "course-authoring.certificates.sidebar.about2.description-1": "To begin issuing course certificates, a course team member with either the Staff or Admin role selects {strongText}. Only course team members with these roles can edit or delete an activated certificate.",
   "course-authoring.certificates.sidebar.about2.description-1.strong": "Activate",
   "course-authoring.certificates.sidebar.about2.description-2": "{strongText} delete certificates after a course has started; learners who have already earned certificates will no longer be able to access them.",
